+++ app/controllers/db/casts_controller.rb
@@ -16,13 +16,13 @@ module Db
 
     def new
       @work = Work.without_deleted.find(params[:work_id])
-      @form = Db::CastRowsForm.new
+      @form = Deprecated::Db::CastRowsForm.new
       authorize @form
     end
 
     def create
       @work = Work.without_deleted.find(params[:work_id])
-      @form = Db::CastRowsForm.new(cast_rows_form_params)
+      @form = Deprecated::Db::CastRowsForm.new(cast_rows_form_params)
       @form.user = current_user
       @form.work = @work
       authorize @form
