+++ app/controllers/db/series_works_controller.rb
@@ -11,13 +11,13 @@ module Db
 
     def new
       @series = Series.without_deleted.find(params[:series_id])
-      @form = Db::SeriesWorkRowsForm.new
+      @form = Deprecated::Db::SeriesWorkRowsForm.new
       authorize @form, :new?
     end
 
     def create
       @series = Series.without_deleted.find(params[:series_id])
-      @form = Db::SeriesWorkRowsForm.new(series_work_rows_form_params)
+      @form = Deprecated::Db::SeriesWorkRowsForm.new(series_work_rows_form_params)
       @form.user = current_user
       @form.series = @series
       authorize @form, :create?
