+++ app/forms/deprecated/db/program_rows_form.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module Db
+module Deprecated::Db
   class ProgramRowsForm
     include ActiveModel::Model
     include Virtus.model
